+++ builtin/describe.c
@@ -236,11 +236,11 @@ static int compare_pt(const void *a_, const void *b_)
 	return 0;
 }
 
-static unsigned long finish_depth_computation(
+static size_t finish_depth_computation(
 	struct commit_list **list,
 	struct possible_tag *best)
 {
-	unsigned long seen_commits = 0;
+	size_t seen_commits = 0;
 	while (*list) {
 		struct commit *c = pop_commit(list);
 		struct commit_list *parents = c->parents;
@@ -306,7 +306,7 @@ static void describe_commit(struct object_id *oid, struct strbuf *dst)
 	struct commit_name *n;
 	struct possible_tag all_matches[MAX_TAGS];
 	unsigned int match_cnt = 0, annotated_cnt = 0, cur_match;
-	unsigned long seen_commits = 0;
+	size_t seen_commits = 0;
 	unsigned int unannotated_cnt = 0;
 
 	cmit = lookup_commit_reference(the_repository, oid);
