+++ http-backend.c
@@ -18,7 +18,7 @@ static const char content_type[] = "Content-Type";
 static const char content_length[] = "Content-Length";
 static const char last_modified[] = "Last-Modified";
 static int getanyfile = 1;
-static unsigned long max_request_buffer = 10 * 1024 * 1024;
+static size_t max_request_buffer = 10 * 1024 * 1024;
 
 static struct string_list *query_params;
 
@@ -372,7 +372,7 @@ static void inflate_request(const char *prog_name, int out, int buffer_input, ss
 	unsigned char *full_request = NULL;
 	unsigned char in_buf[8192];
 	unsigned char out_buf[8192];
-	unsigned long cnt = 0;
+	size_t cnt = 0;
 	int req_len_defined = req_len >= 0;
 	size_t req_remaining_len = req_len;
